Create lang component, simplify t options arg
