+++ lxl-web/src/routes/(app)/[[lang=lang]]/+page.svelte
@@ -4,10 +4,7 @@
 
 <h1>
 	{data.t('home.welcome_text', {
-		values: {
-			site: 'Libris',
-			day: `${(() =>
-				new Intl.DateTimeFormat(data.locale, { weekday: 'long' }).format(new Date()))()}`
-		}
+		site: 'Libris',
+		day: `${(() => new Intl.DateTimeFormat(data.locale, { weekday: 'long' }).format(new Date()))()}`
 	})}
 </h1>
