+++ lxl-web/src/routes/(app)/[[lang=lang]]/search/+page.svelte
@@ -7,7 +7,7 @@
 	$: q = $page.url.searchParams.get('q');
 </script>
 
-<h1>{$page.data.t('search.result_info', { values: { q: `${q}` } })}</h1>
+<h1>{$page.data.t('search.result_info', { q: `${q}` })}</h1>
 <ul>
 	{#each data.items as item (item['@id'])}
 		<li>{item['@id']}</li>
